Separate client componet for useSearchParams and add suspense for fallback when data is loaded.
