+++ src/app/explore/page.tsx
@@ -1,141 +1,16 @@
-'use client'
-
-import FilterBar from '@/modules/explore/components/FilterBar'
+import { Suspense } from 'react'
 import ViewLayout from '@/shared/components/layout/ViewLayout'
 import Container from '@/shared/components/layout/Container'
-import { ScrollArea } from '@/shared/components/ui/base/scroll-area'
-import React, { useEffect, useState } from 'react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import BusinessCards from '@/modules/explore/components/BusinessCards'
-import { FilterKey } from '@/modules/explore/types/filterTypes'
+import ExploreMain from '@/modules/explore/components/ExploreMain'
 
-import Map from '@/modules/explore/components/mapbox/Map'
-import { Business } from '@/modules/explore/types/businessTypes'
-import mockUserBookmarks from '@/shared/mock/user/userBookmarks.json'
-import { useFilters } from '@/modules/explore/hooks/useFilters'
-import { Filters } from '@/modules/explore/types/filterTypes'
-import {
-  fetchFilteredBusinesses,
-  getFiltersFromSearchParams,
-} from '@/modules/explore/services/filterBusinesses'
-import { buildQueryParams } from '@/modules/explore/libs/buildQueryParams'
-
-// import { useSession } from 'next-auth/react'
-// import {
-//   getUserBookmarks,
-//   addUserBookmark,
-//   removeUserBookmark,
-// } from '@/modules/explore/services/bookmark'
-
-interface MockUserBookmark {
-  userId: string
-  businessIds: string[]
-}
-
-const ExplorePage = () => {
-  const [selectedId, setSelectedId] = useState<string | undefined>(undefined)
-  const [hoveredId, setHoveredId] = useState<string | null>(null)
-  // Find the mock user's bookmarked business IDs
-  const mockUserId = 'user1'
-  const initialBookmarkedIds =
-    (mockUserBookmarks as MockUserBookmark[]).find(u => u.userId === mockUserId)?.businessIds || []
-  const [bookmarkedIds, setBookmarkedIds] = useState<string[]>(initialBookmarkedIds)
-  const [businesses, setBusinesses] = useState<Business[]>([])
-
-  // Filter state generated from search params.
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const filters = getFiltersFromSearchParams(searchParams)
-
-  // Hook to get necessary filter options
-  const { filterOptions } = useFilters()
-
-  // Changes in search params trigger a fetch of businesses.
-  useEffect(() => {
-    const filters = getFiltersFromSearchParams(searchParams)
-    fetchFilteredBusinesses(filters).then(({ businesses }) => {
-      setBusinesses(businesses)
-    })
-  }, [searchParams])
-
-  // Uncomment when backend integration is ready
-  // const { data: session } = useSession()
-  // console.log('Session info:')
-  // console.log(session)
-
-  // const [bookmarkedIds, setBookmarkedIds] = useState<string[]>([])
-  // useEffect(() => {
-  //   if (!session?.user?.id) return
-  //   getUserBookmarks(session.user.id)
-  //     .then(ids => setBookmarkedIds(ids))
-  //     .catch(() => setBookmarkedIds([]))
-  // }, [session?.user?.id])
-
-  // Update bookmarks in local state only
-  const handleToggleBookmark = (business_id: string) => {
-    // if (!session?.user?.id) return
-    setBookmarkedIds(prev => {
-      const isBookmarked = prev.includes(business_id)
-      if (isBookmarked) {
-        // removeUserBookmark(session.user.id, business_id)
-        const updated = prev.filter(id => id !== business_id)
-        console.log('Unbookmarked:', business_id, 'Current bookmarks:', updated)
-        return updated
-      } else {
-        // addUserBookmark(session.user.id, business_id)
-        const updated = [...prev, business_id]
-        console.log('Bookmarked:', business_id, 'Current bookmarks:', updated)
-        return updated
-      }
-    })
-  }
-
-  // Update the filter state and push the new search params to the url.
-  const handleFilterChange = <K extends FilterKey>(key: K, value: Filters[K]) => {
-    const nextFilters = { ...filters, [key]: value }
-    const query = buildQueryParams(nextFilters)
-    router.push(query ? `/explore?${query}` : '/explore')
-  }
-
-  // Reset the filters and push the new search params to the url.
-  const resetFilters = () => {
-    router.push('/explore')
-  }
-
+export default function ExplorePage() {
   return (
     <ViewLayout type="explore">
       <Container>
-        <div className="screen-minus-navbar-explore mt-[21px] flex flex-col space-y-8">
-          <FilterBar
-            filters={filters}
-            onChange={handleFilterChange}
-            onReset={resetFilters}
-            filterOptions={filterOptions}
-          />
-          <div className="mb-24 flex min-h-0 flex-1 gap-8">
-            {/* Left: Scrollable business list */}
-            <div className="flex min-h-0 w-[665px] flex-1 flex-col">
-              <ScrollArea className="bg-background h-full min-h-0 flex-1 rounded-xl">
-                <BusinessCards
-                  items={businesses}
-                  onSelect={setSelectedId}
-                  selectedId={selectedId ?? ''}
-                  hoveredId={hoveredId}
-                  onHover={setHoveredId}
-                  bookmarkedIds={bookmarkedIds}
-                  onToggleBookmark={handleToggleBookmark}
-                />
-              </ScrollArea>
-            </div>
-            {/* Right: Mapbox placeholder */}
-            <div className="border-border flex-1 overflow-hidden rounded-xl border">
-              <Map items={businesses} selectedId={selectedId} hoveredId={hoveredId} />
-            </div>
-          </div>
-        </div>
+        <Suspense fallback={<p className="py-20 text-center">Loading...</p>}>
+          <ExploreMain />
+        </Suspense>
       </Container>
     </ViewLayout>
   )
 }
-
-export default ExplorePage
